requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
deepface==0.0.83
python-multipart==0.0.7
pillow==10.3.0
tf-keras==2.19.0
